Export ExportPayload method
